requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-ollama
sentence_transformers
chromadb
langchain_huggingface
faiss-cpu
beautifulsoup4
streamlit
langchain_groq
fastapi
uvicorn
langserve
duckduckgo-search